Two new classes for teleop
Align: Adjust the robot's position and heading until an april tag is at
a specific location in the limelight's POV
TrajectoryLoader: Loads a Swerve trajectory spec from json read either
from NT or from resources. The NT is for easy editing and tweaking of
trajectories in the OutlineViewer without the need to rebuild. Resources
is for more permanent storage and saving of many trajectories.